+++ src/component/user/common/ResetPasswordForm/ResetPasswordForm.tsx
@@ -1,10 +1,4 @@
-import {
-    Button,
-    TextField,
-    IconButton,
-    InputAdornment,
-} from '@material-ui/core';
-import { Visibility, VisibilityOff } from '@material-ui/icons';
+import { Button } from '@material-ui/core';
 import classnames from 'classnames';
 import {
     SyntheticEvent,
@@ -23,6 +17,7 @@ import PasswordMatcher from './PasswordMatcher/PasswordMatcher';
 import { useStyles } from './ResetPasswordForm.styles';
 import { useCallback } from 'react';
 import { formatApiPath } from '../../../../utils/format-path';
+import PasswordField from '../../../common/PasswordField/PasswordField';
 
 interface IResetPasswordProps {
     token: string;
@@ -38,7 +33,6 @@ const ResetPasswordForm = ({ token, setLoading }: IResetPasswordProps) => {
     const [confirmPassword, setConfirmPassword] = useState('');
     const [matchingPasswords, setMatchingPasswords] = useState(false);
     const [validOwaspPassword, setValidOwaspPassword] = useState(false);
-    const [showPassword, setShowPassword] = useState(false);
     const history = useHistory();
 
     const submittable = matchingPasswords && validOwaspPassword;
@@ -46,14 +40,6 @@ const ResetPasswordForm = ({ token, setLoading }: IResetPasswordProps) => {
     const setValidOwaspPasswordMemo = useCallback(setValidOwaspPassword, [
         setValidOwaspPassword,
     ]);
-
-    const handleClickShowPassword = () => {
-        setShowPassword(!showPassword);
-    };
-
-    const handleMouseDownPassword = (e: React.MouseEvent<HTMLButtonElement>) => {
-        e.preventDefault();
-    };
 
     useEffect(() => {
         if (!password) {
@@ -122,64 +108,22 @@ const ResetPasswordForm = ({ token, setLoading }: IResetPasswordProps) => {
                     styles.container
                 )}
             >
-                <TextField
-                    variant="outlined"
-                    size="small"
-                    type={showPassword ? 'text' : 'password'}
-                    InputProps={{
-                        style:{
-                            paddingRight: 0
-                        },
-                        endAdornment: (
-                            <InputAdornment position="end">
-                                <IconButton
-                                    aria-label="toggle password visibility"
-                                    onClick={handleClickShowPassword}
-                                    onMouseDown={handleMouseDownPassword}
-                                >
-                                    {showPassword ? (
-                                        <Visibility />
-                                    ) : (
-                                        <VisibilityOff />
-                                    )}
-                                </IconButton>
-                            </InputAdornment>
-                        ),
-                    }}
+                <PasswordField
                     placeholder="Password"
                     value={password || ''}
-                    onChange={e => setPassword(e.target.value)}
+                    onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
+                        setPassword(e.target.value)
+                    }
                     onFocus={() => setShowPasswordChecker(true)}
                     autoComplete="password"
                     data-loading
                 />
-                <TextField
-                    variant="outlined"
-                    size="small"
-                    type={showPassword ? 'text' : 'password'}
-                    InputProps={{
-                        style:{
-                            paddingRight: 0
-                        },
-                        endAdornment: (
-                            <InputAdornment position="end">
-                                <IconButton
-                                    aria-label="toggle password visibility"
-                                    onClick={handleClickShowPassword}
-                                    onMouseDown={handleMouseDownPassword}
-                                >
-                                    {showPassword ? (
-                                        <Visibility />
-                                    ) : (
-                                        <VisibilityOff />
-                                    )}
-                                </IconButton>
-                            </InputAdornment>
-                        ),
-                    }}
+                <PasswordField
                     value={confirmPassword || ''}
                     placeholder="Confirm password"
-                    onChange={e => setConfirmPassword(e.target.value)}
+                    onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
+                        setConfirmPassword(e.target.value)
+                    }
                     autoComplete="confirm-password"
                     data-loading
                 />
